End of section 17

+++ API/API/SignalR/MessageHub.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.SignalR;
 
 namespace API.SignalR
 {
-    public class MessageHub(IMessageRepository messageRepository, IUserRepository userRepository, IMapper mapper, IHubContext<PesenceHub> presenceHub) : Hub
+    public class MessageHub(IMessageRepository messageRepository, IUserRepository userRepository, IMapper mapper, IHubContext<PresenceHub> presenceHub) : Hub
     {
         public override async Task OnConnectedAsync()
         {
@@ -15,32 +15,37 @@ namespace API.SignalR
             var otherUser = httpContext?.Request.Query["user"];
 
             if (Context.User == null || string.IsNullOrEmpty(otherUser))
-                throw new Exception("Can not join group");
-
+                throw new Exception("Cannot join group");
             var groupName = GetGroupName(Context.User.GetUserName(), otherUser);
             await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
-            await AddToGroup(groupName);
+            var group = await AddToGroup(groupName);
+
+            await Clients.Group(groupName).SendAsync("UpdatedGroup", group);
 
             var messages = await messageRepository.GetMessageThread(Context.User.GetUserName(), otherUser!);
-            await Clients.Group(groupName).SendAsync("ReceiveMessagethred", messages);
+
+            await Clients.Caller.SendAsync("ReceiveMessageThread", messages);
+
         }
+
         public override async Task OnDisconnectedAsync(Exception? exception)
         {
-            await RemoveFromMessageGroup();
+            var group = await RemoveFromMessageGroup();
+            await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
             await base.OnDisconnectedAsync(exception);
         }
 
         public async Task SendMessage(CreateMessageDto createMessageDto)
         {
-            var username = Context.User?.GetUserName() ?? throw new Exception("could not get user.");
+            var username = Context.User?.GetUserName() ?? throw new Exception("could not get user");
 
             if (username == createMessageDto.RecipientUsername.ToLower())
-                throw new HubException("you cannot message yourself");
+                throw new HubException("You cannot message yourself");
 
             var sender = await userRepository.GetUserByUserNameAsync(username);
             var recipient = await userRepository.GetUserByUserNameAsync(createMessageDto.RecipientUsername);
 
-            if (recipient == null || sender == null || recipient.UserName == null || sender.UserName == null)
+            if (recipient == null || sender == null || sender.UserName == null || recipient.UserName == null)
                 throw new HubException("Cannot send message at this time");
 
             var message = new Message
@@ -55,35 +60,35 @@ namespace API.SignalR
             var groupName = GetGroupName(sender.UserName, recipient.UserName);
             var group = await messageRepository.GetMessageGroup(groupName);
 
-            if(group != null && group.Connections.Any(x=>x.UserName == recipient.UserName))
+            if (group != null && group.Connections.Any(x => x.UserName == recipient.UserName))
             {
                 message.DateRead = DateTime.UtcNow;
             }
             else
             {
                 var connections = await PresenceTracker.GetConnectionsForUser(recipient.UserName);
-                if(connections != null && connections?.Count != null){
+                if (connections != null && connections?.Count != null)
+                {
                     await presenceHub.Clients.Clients(connections).SendAsync("NewMessageReceived",
                         new { username = sender.UserName, knownAs = sender.KnownAs });
                 }
             }
 
-                messageRepository.AddMessage(message);
+            messageRepository.AddMessage(message);
 
             if (await messageRepository.SaveAllAsync())
             {
-              
                 await Clients.Group(groupName).SendAsync("NewMessage", mapper.Map<MessageDto>(message));
             }
         }
 
-        private async Task<bool> AddToGroup(string groupName)
+        private async Task<Group> AddToGroup(string groupName)
         {
             var username = Context.User?.GetUserName() ?? throw new Exception("Cannot get username");
             var group = await messageRepository.GetMessageGroup(groupName);
             var connection = new Connection { ConnectionId = Context.ConnectionId, UserName = username };
 
-            if(group == null)
+            if (group == null)
             {
                 group = new Group { Name = groupName };
                 messageRepository.AddGroup(group);
@@ -91,24 +96,28 @@ namespace API.SignalR
 
             group.Connections.Add(connection);
 
-            return await messageRepository.SaveAllAsync();
+            if (await messageRepository.SaveAllAsync()) return group;
+
+            throw new HubException("Failed to join group");
         }
 
-        private async Task RemoveFromMessageGroup()
+        private async Task<Group> RemoveFromMessageGroup()
         {
-            var connection = await messageRepository.GetConnection(Context.ConnectionId);
-            if(connection != null)
+            var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
+            var connection = group?.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
+            if (connection != null && group != null)
             {
                 messageRepository.RemoveConnection(connection);
-                await messageRepository.SaveAllAsync();
+                if (await messageRepository.SaveAllAsync()) return group;
             }
-        }
 
+            throw new Exception("Failed to remove from group");
+        }
 
         private string GetGroupName(string caller, string? other)
         {
-            var stringcompare = string.CompareOrdinal(caller, other) < 0;
-            return stringcompare ? $"{caller}-{other}" : $"{other}-{caller}";
+            var stringCompare = string.CompareOrdinal(caller, other) < 0;
+            return stringCompare ? $"{caller}-{other}" : $"{other}-{caller}";
         }
 
     }
